requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
aniso8601==9.0.1
anyio==3.6.2
appdirs==1.4.4
async-timeout==4.0.2
asyncer==0.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
coloredlogs==15.0.1
cvzone==1.5.6
fastapi==0.87.0
filetype==1.2.0
Flask==2.2.2
Flask-RESTful==0.3.9
flatbuffers==23.1.21
frozenlist==1.3.3
h11==0.14.0
humanfriendly==10.0
idna==3.4
ImageHash==4.3.1
imageio==2.25.0
itsdangerous==2.1.2
Jinja2==3.1.2
llvmlite==0.39.1
MarkupSafe==2.1.2
mpmath==1.2.1
multidict==6.0.4
networkx==3.0
numba==0.56.4
numpy==1.23.5
onnxruntime==1.13.1
opencv-python==4.7.0.68
opencv-python-headless==4.6.0.66
packaging==23.0
Pillow==9.3.0
pooch==1.6.0
protobuf==4.21.12
pydantic==1.10.4
PyMatting==1.1.8
python-multipart==0.0.5
pytz==2022.7.1
PyWavelets==1.4.1
rembg==2.0.30
requests==2.28.2
scikit-image==0.19.3
scipy==1.9.3
six==1.16.0
sniffio==1.3.0
starlette==0.21.0
sympy==1.11.1
tifffile==2023.2.3
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.14
uvicorn==0.20.0
watchdog==2.1.9
Werkzeug==2.2.2
yarl==1.8.2
